+++ lumigator/frontend/src/components/experiments/LExperimentForm.vue
@@ -96,7 +96,12 @@ import LModelCards from '@/components/experiments/LModelCards.vue'
 import { useToast } from 'primevue/usetoast'
 import type { ToastMessageOptions } from 'primevue'
 import type { Dataset } from '@/types/Dataset'
-import type { createExperimentWithWorkflowsPayload } from '@/sdk/experimentsService'
+import {
+  experimentsService,
+  type createExperimentWithWorkflowsPayload,
+} from '@/sdk/experimentsService'
+import type { Model } from '@/types/Model'
+import { workflowsService } from '@/sdk/workflowsService'
 
 const emit = defineEmits(['l-close-form'])
 
@@ -125,6 +130,32 @@ const filteredDatasets = computed(() =>
   datasets.value.filter((dataset) => dataset.ground_truth === true),
 )
 
+/**
+ * Runs an experiment with multiple models.
+ * Each model triggers a respective evaluation job.
+ *
+ * @param {Object} experimentData - The data for the experiment to run.
+ * @returns {Promise<Array>} The results of the experiment.
+ */
+async function createExperimentWithWorkflows(
+  experimentData: createExperimentWithWorkflowsPayload,
+  models: Model[],
+) {
+  // first we create an experiment as a container
+  const { id: experimentId } = await experimentsService.createExperiment(experimentData)
+
+  // then we create a workflow for each model to be attached to the experiment
+  return Promise.all(
+    models.map((model: Model) =>
+      workflowsService.createWorkflow({
+        ...experimentData,
+        experiment_id: experimentId,
+        model: model.uri,
+      }),
+    ),
+  )
+}
+
 async function triggerExperiment() {
   const experimentPayload: createExperimentWithWorkflowsPayload = {
     name: experimentTitle.value,
@@ -133,7 +164,7 @@ async function triggerExperiment() {
     max_samples: maxSamples.value ? maxSamples.value : 0,
     task: 'summarization',
   }
-  const workflows = await experimentStore.createExperimentWithWorkflows(
+  const workflows = await createExperimentWithWorkflows(
     experimentPayload,
     modelSelection.value.selectedModels,
   )
